refactor: selead class 활용으로 가독성 향상시키도록 리팩토링
