+++ src/main/kotlin/omock/model/stonestate/Clear.kt
@@ -1,6 +1,5 @@
-package omock.model.stonestate.unplaced
+package omock.model.stonestate
 
 import omock.model.stone.Stone
-import omock.model.stonestate.unplaced.UnPlaced
 
 class Clear(private val stone: Stone) : UnPlaced(stone)
